+++ moneyplant-ui/projects/dashboards/src/lib/echart-chart-builders/apache-echart-builder.ts
