+++ moneyplant-ui/projects/dashboards/src/lib/echart-chart-builders/area/area-chart-builder.ts
