+++ moneyplant-ui/projects/dashboards/src/lib/echart-chart-builders/area/index.ts
