+++ moneyplant-ui/projects/dashboards/src/lib/echart-chart-builders/bar/bar-chart-builder.ts
