+++ moneyplant-ui/projects/dashboards/src/lib/echart-chart-builders/bar/index.ts
