+++ moneyplant-ui/projects/dashboards/src/lib/echart-chart-builders/gauge/gauge-chart-builder.ts
