+++ moneyplant-ui/projects/dashboards/src/lib/echart-chart-builders/gauge/index.ts
