+++ moneyplant-ui/projects/dashboards/src/lib/echart-chart-builders/scatter/index.ts
