+++ moneyplant-ui/projects/dashboards/src/lib/echart-chart-builders/scatter/scatter-chart-builder.ts
