+++ moneyplant-ui/projects/dashboards/src/lib/echart-chart-builders/stacked-area/README-stacked-area.md
