+++ moneyplant-ui/projects/dashboards/src/lib/echart-chart-builders/stacked-area/index.ts
