+++ moneyplant-ui/projects/dashboards/src/lib/echart-chart-builders/stacked-area/stacked-area-chart-builder.ts
