+++ moneyplant-ui/projects/dashboards/src/lib/echart-chart-builders/treemap/README-treemap.md
